Refactor MerakiDevice model into specialized domain models
- Created MerakiBaseDevice in core/models/base.py to hold common fields.
- Decomposed MerakiDevice in core/models/device.py to inherit from MerakiBaseDevice.
- Added specialized device classes (MerakiApplianceDevice, MerakiSwitchDevice, etc.) in their respective domain model files.
- Updated core/models/__init__.py to export new specialized models.
- Refactored sensor components (switch_poe, appliance_uplink, camera_analytics) to use specialized models instead of the aggregate MerakiDevice, reducing context pressure and improving modularity.
- Fixed a broken import in api/websocket.py.

Note: Encountered environment-related import errors during test execution (ModuleNotFoundError for sensor.coordinators) which appear to be pre-existing or related to the repository's structure for relative imports. Completed the architectural decomposition as requested.
