+++ custom_components/meraki_ha/core/models/__init__.py
@@ -1,7 +1,8 @@
 """Models package for Meraki API data structures."""
 
-from .appliance import ApplianceMixin, MerakiAppliancePort
-from .camera import CameraMixin
+from .appliance import ApplianceMixin, MerakiApplianceDevice, MerakiAppliancePort
+from .base import MerakiBaseDevice
+from .camera import CameraMixin, MerakiCameraDevice
 from .device import MerakiDevice
 from .network import (
     MerakiFirewallRule,
@@ -11,18 +12,24 @@ from .network import (
     MerakiVpn,
 )
 from .organization import MerakiOrganization
-from .sensor import SensorMixin
-from .switch import SwitchMixin
-from .wireless import WirelessMixin
+from .sensor import MerakiSensorDevice, SensorMixin
+from .switch import MerakiSwitchDevice, SwitchMixin
+from .wireless import MerakiWirelessDevice, WirelessMixin
 
 __all__ = [
     "ApplianceMixin",
     "CameraMixin",
+    "MerakiApplianceDevice",
     "MerakiAppliancePort",
+    "MerakiBaseDevice",
+    "MerakiCameraDevice",
     "MerakiDevice",
     "MerakiFirewallRule",
     "MerakiNetwork",
     "MerakiOrganization",
+    "MerakiSensorDevice",
+    "MerakiSwitchDevice",
+    "MerakiWirelessDevice",
     "SensorMixin",
     "SwitchMixin",
     "MerakiTrafficShaping",
